refactor: Project structure
Created more organized/convinient directory  structure
and added some organizatives comments
