+++ src/components/main/App.js
@@ -1,5 +1,9 @@
 import React from 'react';
+
+// assets
 import logo from './logo.svg';
+
+// styles
 import './App.css';
 
 function App() {
