+++ src/index.js
@@ -1,7 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+
+// main app dependencies
+import './components/main/index.css';
+import App from './components/main/App';
+
+// service worker
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
